+++ src/cpp/benders/benders_core/include/antares-xpansion/benders/benders_core/CriterionComputation.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "OuterLoopInputDataReader.h"
+#include "CriterionInputDataReader.h"
 #include "VariablesGroup.h"
 namespace Benders::Criterion {
