+++ src/cpp/benders/benders_core/include/antares-xpansion/benders/benders_core/VariablesGroup.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "OuterLoopInputDataReader.h"
+#include "CriterionInputDataReader.h"
 namespace Benders::Criterion {
 class VariablesGroup {
  public:
